refactor: request 패키지 위치 수정 #76
