+++ domain/src/main/java/com/kdjj/domain/model/request/FetchRemoteLatestRecipeListRequest.kt
@@ -1,4 +1,4 @@
-package com.kdjj.domain.request
+package com.kdjj.domain.model.request
 
 data class FetchRemoteLatestRecipeListRequest(
     val lastVisibleCreateTime: Long,
