+++ domain/src/main/java/com/kdjj/domain/model/request/FetchRemoteSearchRecipeListRequest.kt
@@ -1,4 +1,4 @@
-package com.kdjj.domain.request
+package com.kdjj.domain.model.request
 
 data class FetchRemoteSearchRecipeListRequest(
     val keyword: String,
