+++ domain/src/main/java/com/kdjj/domain/model/request/Request.kt
@@ -1,0 +1,3 @@
+package com.kdjj.domain.model.request
+
+interface Request
