--- domain/src/main/java/com/kdjj/domain/request/EmptyRequest.kt
@@ -1,3 +1,0 @@
-package com.kdjj.domain.request
-
-object EmptyRequest : Request
