--- domain/src/main/java/com/kdjj/domain/request/Request.kt
@@ -1,3 +1,0 @@
-package com.kdjj.domain.request
-
-interface Request
