+++ domain/src/main/java/com/kdjj/domain/usecase/DeleteRemoteRecipeUseCase.kt
@@ -1,7 +1,7 @@
 package com.kdjj.domain.usecase
 
 import com.kdjj.domain.repository.RecipeRepository
-import com.kdjj.domain.request.DeleteRemoteRecipeRequest
+import com.kdjj.domain.model.request.DeleteRemoteRecipeRequest
 import javax.inject.Inject
 
 class DeleteRemoteRecipeUseCase @Inject constructor(
