+++ domain/src/main/java/com/kdjj/domain/usecase/FetchLocalFavoriteRecipeListUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.repository.RecipeListRepository
-import com.kdjj.domain.request.FetchLocalFavoriteRecipeListRequest
+import com.kdjj.domain.model.request.FetchLocalFavoriteRecipeListRequest
 import javax.inject.Inject
 
 internal class FetchLocalFavoriteRecipeListUseCase @Inject constructor(
