+++ domain/src/main/java/com/kdjj/domain/usecase/FetchLocalLatestRecipeListUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.repository.RecipeListRepository
-import com.kdjj.domain.request.FetchLocalLatestRecipeListRequest
+import com.kdjj.domain.model.request.FetchLocalLatestRecipeListRequest
 import javax.inject.Inject
 
 internal class FetchLocalLatestRecipeListUseCase @Inject constructor(
