+++ domain/src/main/java/com/kdjj/domain/usecase/FetchLocalSearchRecipeListUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.repository.RecipeListRepository
-import com.kdjj.domain.request.FetchLocalSearchRecipeListRequest
+import com.kdjj.domain.model.request.FetchLocalSearchRecipeListRequest
 import javax.inject.Inject
 
 internal class FetchLocalSearchRecipeListUseCase @Inject constructor(
