+++ domain/src/main/java/com/kdjj/domain/usecase/FetchRecipeTypesUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.RecipeType
 import com.kdjj.domain.repository.RecipeTypeRepository
-import com.kdjj.domain.request.EmptyRequest
+import com.kdjj.domain.model.request.EmptyRequest
 import javax.inject.Inject
 
 class FetchRecipeTypesUseCase @Inject constructor(
