+++ domain/src/main/java/com/kdjj/domain/usecase/FetchRemoteLatestRecipeListUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.repository.RecipeListRepository
-import com.kdjj.domain.request.FetchRemoteLatestRecipeListRequest
+import com.kdjj.domain.model.request.FetchRemoteLatestRecipeListRequest
 import javax.inject.Inject
 
 internal class FetchRemoteLatestRecipeListUseCase @Inject constructor(
