+++ domain/src/main/java/com/kdjj/domain/usecase/FetchRemotePopularRecipeListUseCase.kt
@@ -2,7 +2,7 @@ package com.kdjj.domain.usecase
 
 import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.repository.RecipeListRepository
-import com.kdjj.domain.request.FetchRemotePopularRecipeListRequest
+import com.kdjj.domain.model.request.FetchRemotePopularRecipeListRequest
 import javax.inject.Inject
 
 internal class FetchRemotePopularRecipeListUseCase @Inject constructor(
