+++ domain/src/main/java/com/kdjj/domain/usecase/UseCase.kt
@@ -1,6 +1,6 @@
 package com.kdjj.domain.usecase
 
-import com.kdjj.domain.request.Request
+import com.kdjj.domain.model.request.Request
 
 interface UseCase<R: Request, T> {
     suspend operator fun invoke(request: R): Result<T>
